haunted-house-api/pyproject.toml
[project]
name = "haunted-house-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "supabase>=2.28.0",
    "uvicorn>=0.39.0",
]
